+++ Verizon/src/main/java/com/mopub/mobileads/VerizonRewardedVideo.java
@@ -9,12 +9,14 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.mopub.common.BaseLifecycleListener;
+import com.mopub.common.DataKeys;
 import com.mopub.common.LifecycleListener;
 import com.mopub.common.MoPub;
 import com.mopub.common.MoPubReward;
 import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
 import com.verizon.ads.ActivityStateManager;
+import com.verizon.ads.Bid;
 import com.verizon.ads.CreativeInfo;
 import com.verizon.ads.ErrorInfo;
 import com.verizon.ads.RequestMetadata;
@@ -23,6 +25,7 @@ import com.verizon.ads.edition.StandardEdition;
 import com.verizon.ads.interstitialplacement.InterstitialAd;
 import com.verizon.ads.interstitialplacement.InterstitialAdFactory;
 
+import java.util.HashMap;
 import java.util.Map;
 
 import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
@@ -153,6 +156,8 @@ public class VerizonRewardedVideo extends BaseAd {
         Preconditions.checkNotNull(context);
         Preconditions.checkNotNull(adData);
 
+        final Map<String, String> extras = adData.getExtras();
+
         this.activity = (Activity) context;
 
         VASAds.setLocationEnabled(MoPub.getLocationAwareness() != MoPub.LocationAwareness.DISABLED);
@@ -160,12 +165,29 @@ public class VerizonRewardedVideo extends BaseAd {
         final InterstitialAdFactory interstitialAdFactory = new InterstitialAdFactory(activity,
                 placementId, new VerizonInterstitialFactoryListener());
 
-        final RequestMetadata requestMetadata = new RequestMetadata.Builder().setMediator
-                (VerizonAdapterConfiguration.MEDIATOR_ID)
-                .build();
+        final Bid bid = BidCache.get(placementId);
 
-        interstitialAdFactory.setRequestMetaData(requestMetadata);
-        interstitialAdFactory.load(new VerizonInterstitialListener());
+        if (bid == null) {
+            final RequestMetadata.Builder requestMetadataBuilder = new RequestMetadata.Builder(
+                    VASAds.getRequestMetadata());
+            requestMetadataBuilder.setMediator(VerizonAdapterConfiguration.MEDIATOR_ID);
+
+            final String adContent = extras.get(DataKeys.ADM_KEY);
+
+            if (!TextUtils.isEmpty(adContent)) {
+                final Map<String, Object> placementData = new HashMap<>();
+
+                placementData.put(VerizonAdapterConfiguration.REQUEST_METADATA_AD_CONTENT_KEY, adContent);
+                placementData.put("overrideWaterfallProvider", "waterfallprovider/sideloading");
+
+                requestMetadataBuilder.setPlacementData(placementData);
+            }
+
+            interstitialAdFactory.setRequestMetaData(requestMetadataBuilder.build());
+            interstitialAdFactory.load(new VerizonInterstitialListener());
+        } else {
+            interstitialAdFactory.load(bid, new VerizonInterstitialListener());
+        }
 
         MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
     }